SceneLoader스크립트를 제거한다
SceneLoader스크립트는 딱 한 군데에서만 사용하고, 그 사용처조차도 넓게 확장되지는 않기 때문에
SceneLoader스크립트를 제거한다.

기존의 SceneLoader가 했던 역할은 MainScene에게 이관하며(SceneLoader스크립트는 MainScene씬에서만 사용됨)
ContentBlock에게 '선택된 상태에서 클릭할때의 이벤트'를 추가하여 기존과 마찬가지로 작동하게끔 한다.
