+++ Assets/Scenes/MainScene.unity
@@ -1005,7 +1005,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1364285833}
-  - component: {fileID: 1364285835}
   - component: {fileID: 1364285836}
   - component: {fileID: 1364285837}
   m_Layer: 0
@@ -1029,20 +1028,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1364285835
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1364285832}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b91e0e6bf62ea8644b9913c99a154e53, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  _ScrollVlew: {fileID: 1831129426}
-  _ExitText: {fileID: 1177533451}
 --- !u!114 &1364285836
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1072,6 +1057,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9a9f674ded444f64085ab649095d57c6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  _ExitText: {fileID: 1177533451}
+  _ScrollVlew: {fileID: 1831129426}
 --- !u!114 &1382144570 stripped
 MonoBehaviour:
   m_CorrespondingSourceObject: {fileID: 4080945149712710271, guid: 8e9bea0c42c59cb4aa329f6490d1e902,
