--- Assets/Script/MainScene/SceneLoader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b91e0e6bf62ea8644b9913c99a154e53
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
